Generator moved
I decided to also follow the Pastboard dependency and get rid of the specific file just for the generator.
